modified blob

+++ PlayHarmoniez/Controllers/SongController.cs
@@ -58,12 +58,11 @@ namespace PlayHarmoniez.Controllers
                     Description = song.Description,
                     AlbumId = song.AlbumId,
                     Album = song.Album,
+                    SoundFile = await UploadSong(song.Title, musicContainerName, soundFile),
+                    ImageFile = await UploadImage(song.Title, musicImageContainerName, imageFile),
                     PlaylistSongs = song.PlaylistSongs,
                     LikedSong = song.LikedSong
                 };
-
-                songModel.SoundFile = await UploadSong(songModel.Title, musicContainerName, soundFile);
-                songModel.ImageFile = await UploadImage(songModel.Title, musicImageContainerName, imageFile);
 
                 await _dataContext
                     .Songs
@@ -122,14 +121,11 @@ namespace PlayHarmoniez.Controllers
                     Description = song.Description,
                     AlbumId = song.AlbumId,
                     Album = song.Album,
-                    SoundFile = song.SoundFile,
-                    ImageFile = song.ImageFile,
+                    SoundFile = await UploadSong(song.Title, musicContainerName, soundFile),
+                    ImageFile = await UploadImage(song.Title, musicImageContainerName, imageFile),
                     PlaylistSongs = song.PlaylistSongs,
                     LikedSong = song.LikedSong,
                 };
-
-                updatedSong.SoundFile = await UploadSong(updatedSong.Title, musicContainerName, soundFile);
-                updatedSong.ImageFile = await UploadImage(updatedSong.Title, musicImageContainerName, imageFile);
 
                 return View(updatedSong);
             }
@@ -148,8 +144,8 @@ namespace PlayHarmoniez.Controllers
                     songModel.AlbumId = song.AlbumId;
                     songModel.Album = song.Album;songModel.PlaylistSongs = song.PlaylistSongs;
                     songModel.LikedSong = song.LikedSong;
-                    songModel.SoundFile = await UploadSong(songModel.Title, musicContainerName, soundFile);
-                    songModel.ImageFile = await UploadImage(songModel.Title, musicImageContainerName, imageFile);
+                    songModel.SoundFile = await UploadSong(song.Title, musicContainerName, soundFile);
+                    songModel.ImageFile = await UploadImage(song.Title, musicImageContainerName, imageFile);
 
                     await _dataContext.SaveChangesAsync();
 
@@ -175,19 +171,25 @@ namespace PlayHarmoniez.Controllers
 
                 var song = await _dataContext.Songs.FindAsync(id);
 
-                var containerClient = _blobClient.GetBlobContainerClient(musicContainerName);
-                var blobClient = containerClient.GetBlobClient(song.Title);
+                if (song != null)
+                {
+                    // deleting the sound file
+                    var containerClient = _blobClient.GetBlobContainerClient(musicContainerName);
+                    var blobClient = containerClient.GetBlobClient(song.Title);
 
-                await blobClient.DeleteIfExistsAsync();
+                    await blobClient.DeleteIfExistsAsync();
 
-                containerClient = _blobClient.GetBlobContainerClient(musicImageContainerName);
-                blobClient = containerClient.GetBlobClient(song.Title);
+                    containerClient = _blobClient.GetBlobContainerClient(musicImageContainerName);
 
-                await blobClient.DeleteIfExistsAsync();
+                    // deleting the song image file
+                    blobClient = containerClient.GetBlobClient(song.Title);
 
-                _dataContext.Songs.Remove(song);
+                    await blobClient.DeleteIfExistsAsync();
 
-                await _dataContext.SaveChangesAsync();
+                    _dataContext.Songs.Remove(song);
+
+                    await _dataContext.SaveChangesAsync();
+                }
 
                 return RedirectToAction("SongsList");
             }
@@ -209,6 +211,7 @@ namespace PlayHarmoniez.Controllers
                         ImageFile = songModel.ImageFile,
                     };
                 }
+
                 return null;
             }
